+++ src/odr/internal/html/document.cpp
@@ -54,9 +54,12 @@ void front(const Document &document, const WritingState &state,
                           : std::nullopt);
 
   write_document_style(state);
+  write_document_dark_style(state);
   write_search_style(state);
+  write_search_dark_style(state);
   if (document.document_type() == DocumentType::spreadsheet) {
     write_spreadsheet_style(state);
+    write_spreadsheet_dark_style(state);
   }
 
   out.write_header_end();
@@ -317,7 +320,8 @@ public:
       out.write_element_end("div");
       out.write_element_end("div");
     } else {
-      out.write_element_begin("div");
+      out.write_element_begin("div",
+                              HtmlElementOptions().set_class("odr-text-flow"));
       translate_children(element.children(), state);
       out.write_element_end("div");
     }